Refactor out config part
