+++ pkg/function/create_volume_from_snapshot.go
@@ -11,7 +11,7 @@ import (
 
 	kanister "github.com/kanisterio/kanister/pkg"
 	"github.com/kanisterio/kanister/pkg/blockstorage"
-	"github.com/kanisterio/kanister/pkg/blockstorage/awsebs"
+	awsconfig "github.com/kanisterio/kanister/pkg/config/aws"
 	"github.com/kanisterio/kanister/pkg/blockstorage/getter"
 	"github.com/kanisterio/kanister/pkg/kube"
 	kubevolume "github.com/kanisterio/kanister/pkg/kube/volume"
@@ -60,9 +60,9 @@ func createVolumeFromSnapshot(ctx context.Context, cli kubernetes.Interface, nam
 		}
 		switch pvcInfo.Type {
 		case blockstorage.TypeEBS:
-			config[awsebs.ConfigRegion] = pvcInfo.Region
-			config[awsebs.AccessKeyID] = profile.Credential.KeyPair.ID
-			config[awsebs.SecretAccessKey] = profile.Credential.KeyPair.Secret
+			config[awsconfig.ConfigRegion] = pvcInfo.Region
+			config[awsconfig.AccessKeyID] = profile.Credential.KeyPair.ID
+			config[awsconfig.SecretAccessKey] = profile.Credential.KeyPair.Secret
 		case blockstorage.TypeGPD:
 			config[blockstorage.GoogleProjectID] = profile.Credential.KeyPair.ID
 			config[blockstorage.GoogleServiceKey] = profile.Credential.KeyPair.Secret
